requirements.txt
openai
python-dotenv
langchain
streamlit
transforms
torch